Add devcontainer
It may often be undesirable to have to manually install these tools on a
per-project basis; and so this installs the tools into a basic
devcontainer that is used by the project. This simplifies the
development experience by providing all the tooling and VSCode
extensions required for the Sushi project, without forcing users to have
direct access to the tools, or to NPM, itself.